feat: authentication and authorization added several databases and authentication and authorization first layout for implementation. notification and other supportive elements will also be updated and migrated in the next commit. I have to fix the way i commit stuff.

+++ athena/lib/athena/assessments/quiz_question.ex
@@ -5,7 +5,7 @@ defmodule Athena.Assessments.QuizQuestion do
   schema "quiz_questions" do
     field(:question_text, :string)
     field(:question_type, Ecto.Enum, values: [:choice, :writing], default: :choice)
-    field(:points, :float)
+    field(:points, :decimal)
     field(:order_index, :integer)
 
     belongs_to(:quiz, Athena.Assessments.Quiz)
